+++ apps/client/public/CHANGELOG.md
@@ -16,6 +16,7 @@ If you found some bugs or want us to add new feature, please do so via [our GitH
 - Added checking for duplicated enBID + CLID in LTE
 - Added `editor` role
 - Calculate the speed of radiolines and show total speed
+- Added more checks on the server for duplicates per operator etc
 
 # 2026-02-21